% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survaux-data.R
\docType{data}
\name{ACTG175}
\alias{ACTG175}
\title{ACTG175 - AIDS Clinical Trials Group Study 175}
\format{
This data frame contains the following columns:

\describe{
\item{pidnum}{patient's ID number}
\item{age}{age in years at baseline}
}
}
\source{
Hammer SM, et al. (1996), "A trial comparing nucleoside monotherapy
with combination therapy in HIV-infected adults with CD4 cell counts from
200 to 500 per cubic millimeter.", New England Journal of Medicine, 335:1081–1090.
}
\usage{
ACTG175
}
\description{
ACTG 175 was a randomized clinical trial to compare monotherapy
with zidovudine or didanosine with combination therapy with zidovudine and
didanosine or zidovudine and zalcitabine in adults infected with the human
immunodeficiency virus type I whose CD4 T cell counts were between 200 and
500 per cubic millimeter.
}
\details{
The variable days contains right-censored time-to-event observations.
The data set includes the following post-randomization covariates: CD4 and CD8
T cell count at 20\\pm5 weeks and the indicator of whether or not the patient
was taken off-treatment before 96\\pm5 weeks.
}
\examples{

data(ACTG175)
yobs  <- ACTG175$days / 365
delta <- ACTG175$cens
X     <- as.matrix( cbind(ACTG175['cd40']/100,
                    ACTG175['treat']) ) # summary(X0)

}
\keyword{datasets}
