% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_AR_AGMM.R
\name{AR.AuxSP.AGMM}
\alias{AR.AuxSP.AGMM}
\alias{AR.AuxSP.AGMM.fit}
\alias{AR.AuxSP.AGMM.Loss}
\alias{AR.AuxSP.AGMM.Profile}
\alias{AR.AuxSP.AGMM.Initial}
\title{Fit additive risk model with auxiliary subgroup survival rates using AGMM}
\usage{
AR.AuxSP.AGMM(
  formula,
  data,
  aux,
  na.action = na.omit,
  lambdas = NULL,
  lambdas.num = 10,
  threshold = 1e-04,
  Profile = FALSE,
  trace.cv = TRUE
)
}
\arguments{
\item{formula}{formula object, with the response on the left of a ~ operator,
and the terms on the right. The response must be a survival object as
returned by the Surv function.}

\item{data}{a data.frame in which to interpret the variables named in the formula.}

\item{aux}{indicates the historical summary data. It is a list of lists, and each sub-list represents auxiliary information from a study.
In each study, we have several time points and each time point contains the following four elements
\code{tstar} the time point that the auxiliary information was calculated at;
\code{gfun} a function used to identify the subgroup;
\code{sprob} auxiliary subgroup survival rates for each subgroup at the current time point.}

\item{na.action}{a missing-data filter function.
This is applied to the model.frame after any subset argument has been used.}

\item{lambdas}{user-specified tuning paramters.
default is NULL.}

\item{lambdas.num}{number tuning parameters that will be evaluated at.
default is 10.}

\item{threshold}{a threshold. The absolute values of estimates of tau smaller than the threshold are treated as 0.}

\item{Profile}{a logical value. Whether use the profiled algorithm or not.}

\item{trace.cv}{a logical value. Print the calculating process or not}
}
\value{
A list \code{out} representing the fit,
which contains the following elements:
\code{coef} for estimated coefficients,
\code{convergence} for the optimizatoin is converged or not.
}
\description{
This is a function used to fit the additive risk model
in the presence of potentially incomparable subgroup survival rates
as auxiliary information based on the GMM framework and penalization.
The method is motivated by Chen et al. (2020).
}
\examples{
## import the data:
data("toydata1")
## prepare the auxiliary information:
aux <- list(
  study1 = list(
    time1 = list(
      tstar=0.5,
      sprob=c(0.646,0.732,0.433,0.491)-c(0.2,0,0,0),
      gfunc=function(X){  # group function
        rbind(  ( X[,1] >= 0.5 & X[,2] == 0),
                ( X[,1] <  0.5 & X[,2] == 0),
                ( X[,1] >= 0.5 & X[,2] == 1),
                ( X[,1] <  0.5 & X[,2] == 1)
        ) * 1
      }
    )
  )
)
## fit the model:
sol.AGMM <- AR.AuxSP.AGMM(Surv(yobs,delta)~X1+X2,data=toydata1,aux=aux,lambdas.num=5,trace.cv=TRUE)

}
\references{
Chen, Z., Ning, J., Shen, Y., and Qin, J. (2020).
Combining primary cohort data with external aggregate information
without assuming comparability.
Biometrics. 77(3): 1024-1036.
}
