% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AR.AuxSP.GMM.R
\name{AR.AuxSP.EE}
\alias{AR.AuxSP.EE}
\alias{AR.AuxSP.EE.Sigma}
\title{subgroup survival rate estimating equation}
\usage{
AR.AuxSP.EE(bet, alp, yobs, delta, X, tstar, phi, G)
}
\arguments{
\item{bet}{betas that will be computed at}

\item{alp}{alpha that will be computed at}

\item{tstar}{auxiliary time point (only 1 is allowed).}

\item{phi}{auxiliary survival rates (length(phi)=nrow(G)).}

\item{G}{indicator matrix - one row indicates one group.}
}
\value{
A vector \code{Psin}.
}
\description{
This is a function used to calculate the estimating equation
for the subgroup survival rates at t* under the additive risk model (
its derivation is based on 'Double Expectation')
}
\examples{
1
}
