% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AR.AuxSP.GMM.R
\name{AR.AuxSP.GMM}
\alias{AR.AuxSP.GMM}
\alias{AR.AuxSP.GMM.fit}
\alias{AR.AuxSP.GMM.LossFunc}
\alias{AR.AuxSP.GMM.SE}
\title{Fit additive risk model with auxiliary subgroup survival rates using GMM}
\usage{
AR.AuxSP.GMM(formula, data, aux, na.action = na.omit)
}
\arguments{
\item{formula}{formula object, with the response on the left of a ~ operator,
and the terms on the right. The response must be a survival object as
returned by the Surv function.}

\item{data}{a data.frame in which to interpret the variables named in the formula.}

\item{aux}{a list that should contain the auxiliary subgroup survival information.
It has three elements:
\code{tstar} auxiliary time point (only 1 is allowed);
\code{phi} auxiliary survival rates;
\code{G} indicator matrix whose one row indicates one group.
Note that we should let length(phi)=nrow(G).}

\item{na.action}{a missing-data filter function.
This is applied to the model.frame after any subset argument has been used.}
}
\value{
A list \code{out} representing the fit,
which contains:
\code{coef} for estimated coefficients,
\code{convergence} for the optimizatoin is converged or not,
et al..
}
\description{
This is a function used to fit the additive risk model
in the presence of subgroup survival rates as auxiliary information
based on the GMM framework.
}
\examples{
## import the data:
data("toydata1")
## prepare the auxiliary information:
gXcut <- function(X){  # group function
rbind(  ( X[,1] >= 0.5 & X[,2] == 0),
        ( X[,1] <  0.5 & X[,2] == 0),
        ( X[,1] >= 0.5 & X[,2] == 1),
        ( X[,1] <  0.5 & X[,2] == 1)
) * 1
}
aux <- list(tstar=0.5,phi=c(0.646,0.732,0.433,0.491),
            G=gXcut(toydata1[,c("X1","X2")]))
## fit the model:
AR.AuxSP.GMM(Surv(yobs,delta)~X1+X2,data=toydata1,aux=aux)

}
\references{
Shang, W. and Wang, X. (2017).
The generalized moment estimation of the additive–multiplicative hazard model
with auxiliary survival information.
\emph{Computational Statistics & Data Analysis},  \bold{112}:154–169.
}
