% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_AR_LY.R
\name{AR.Ht}
\alias{AR.Ht}
\title{baseline cumulative hazard function in Additive Risk Model}
\usage{
AR.Ht(object, tm = NULL)
}
\arguments{
\item{object}{The result of an AR fit. Like: AR(), AR.AuxSP.GMM(),
AR.AuxSP.AGMM() ....}

\item{tm}{time points that will be calculated at (length m)
If 'NULL', calculate at yobs. (with sort!!!)}
}
\value{
A list \code{out}.
}
\description{
This is a function used to get baseline cumulative hazard
function in Lin and Ying (1994).
}
\examples{
AR(1,1)

}
