% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AR.R
\name{AR}
\alias{AR}
\alias{AR.fit}
\title{Fit Additive Risk Model}
\usage{
AR(formula, data, na.action = na.omit)
}
\arguments{
\item{formula}{formula object, with the response on the left of a ~ operator,
and the terms on the right. The response must be a survival object as
returned by the Surv function.}

\item{data}{a data.frame in which to interpret the variables named in the formula.}

\item{na.action}{a missing-data filter function.
This is applied to the model.frame after any subset argument has been used.}
}
\value{
A list \code{out} representing the fit,
which contains:
\code{coef} for estimated coefficients,
\code{convergence} for the optimizatoin is converged or not,
et al.
}
\description{
This is a function used to fit the additive risk model of Aalen (1989) based on
the estimation procedure proposed by Lin and Ying (1994).
}
\examples{
data("toydata1")
AR(Surv(yobs,delta)~X1+X2,data=toydata1)

}
\references{
Aalen, O. O. (1989).
A linear regression model for the analysis of life times.
Statistics in Medicine, 8(8):907–925.

Lin, D. Y. and Ying, Z. (1994).
Semiparametric analysis of the additive risk model.
Biometrika, 81(1):61–71.
}
