% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AR.R
\name{AR.Stx}
\alias{AR.Stx}
\title{baseline cumulative hazard function in Additive Risk Model}
\usage{
AR.Stx(object, tm = NULL, x = NULL)
}
\arguments{
\item{object}{The result of an AR fit. Like: AR(), AR.AuxSP.GMM(),
AR.AuxSP.AGMM() ....}

\item{tm}{time points that will be calculated at (length m)
If 'NULL', calculate at yobs. (without sort!!!)}

\item{x}{covariates that will be calculated at (length l)
If 'NULL', calculate at X}
}
\value{
A list \code{out}.
}
\description{
This is a function used to get baseline cumulative hazard
function in Lin and Ying (1994).
}
\examples{
AR(1,1)

}
