% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KM.St.Sub.R
\name{KM.St.Sub}
\alias{KM.St.Sub}
\title{Get subgroup survival rates using KM}
\usage{
KM.St.Sub(tstar, yobs, delta, G)
}
\arguments{
\item{yobs}{observed failure time.}

\item{delta}{censoring indicator.}

\item{G}{indicator matrix - one row indicates one group.}

\item{X}{covariates.}
}
\value{
return the subgroup survival rates \code{tSP}, which is
a vector of length \code{nrow(G)}.
}
\description{
This is a function used to calculate survival rates
in each subgroup using KM.
}
\examples{
1
}
