% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_Other.R
\name{AR.AuxSP.EE}
\alias{AR.AuxSP.EE}
\alias{AR.AuxSP.EE.Sigma}
\title{subgroup survival rate estimating equation}
\usage{
AR.AuxSP.EE(bet, alp, yobs, delta, X, auxinfo)
}
\arguments{
\item{bet}{betas that will be computed at}

\item{alp}{alpha that will be computed at}

\item{auxinfo}{auxiliary ifnormation}
}
\value{
A vector \code{Psin}.
}
\description{
This is a function used to calculate the estimating equation
for the subgroup survival rates at t* under the additive risk model (
its derivation is based on 'Double Expectation')
}
\examples{
1
}
