% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AR.AuxSP.EL.R
\name{AR.AuxSP.EL}
\alias{AR.AuxSP.EL}
\title{Fit additive risk model with auxiliary subgroup survival rates using EL}
\usage{
AR.AuxSP.EL(yobs, delta, X, tstar, phi, G, maxit = 30, tol = 1e-05)
}
\arguments{
\item{tstar}{auxiliary time point (only 1 is allowed).}

\item{phi}{auxiliary survival rates (length(phi)=nrow(G)).}

\item{G}{indicator matrix - one row indicates one group.}
}
\value{
A list \code{out} representing the fit,
which contains the following elements:
\code{coef} for estimated coefficients,
\code{convergence} for the optimizatoin is converged or not.
}
\description{
This is a function used to fit the additive risk model
in the presence of subgroup survival rates as auxiliary information
based on the empirical likelihood framework.
}
\examples{
1
}
\references{
He, J., Li, H., Zhang, S., and Duan, X. (2019).
Additive hazards model with auxiliary subgroup survival information.
Lifetime Data Analysis, 25(1):128–149.
}
